+++ examples/vite/whitelist.ts
@@ -1,6 +1,5 @@
-import { Whitelist } from "@polkadot-api/cli"
-import { wnd } from "@polkadot-api/descriptors"
+import { WndWhitelistEntry } from "@polkadot-api/descriptors"
 
-export const whitelist: Whitelist<typeof wnd> = [
+export const whitelist: WndWhitelistEntry[] = [
   "tx.Balances.transfer_keep_alive",
 ]