+++ packages/cli/src/index.ts
@@ -1,3 +1,2 @@
-export type { Whitelist } from "@polkadot-api/codegen"
 export * from "./commands"
 export * from "./cli"